requirements.txt
amqp==2.2.2
anyjson==0.3.3
billiard==3.5.0.3
celery==4.1.0
dj-database-url==0.4.2
Django==1.11.7
django-audit-log==0.7.0
django-countries==5.0
django-crispy-forms==1.7.0
django-datetimepicker==3.14
django-multi-form-view==2.0.0
gunicorn==19.7.1
kombu==4.1.0
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
redis==2.10.6
six==1.10.0
trans==2.1.0
vine==1.1.4
